+++ sdk/storage/azure_storage_blob/tests/blob_client.rs
@@ -122,12 +122,7 @@ async fn test_upload_blob(ctx: TestContext) -> Result<(), Box<dyn Error>> {
 
     // No Overwrite Scenario
     blob_client
-        .upload(
-            RequestContent::from(data.to_vec()),
-            false,
-            u64::try_from(data.len())?,
-            None,
-        )
+        .upload(RequestContent::from(data.to_vec()), None)
         .await?;
 
     // Assert
@@ -148,9 +143,7 @@ async fn test_upload_blob(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     let response = blob_client
         .upload(
             RequestContent::from(new_data.to_vec()),
-            false,
-            u64::try_from(new_data.len())?,
-            None,
+            Some(BlockBlobClientUploadOptions::default().with_if_not_exists()),
         )
         .await;
 
@@ -160,19 +153,13 @@ async fn test_upload_blob(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     assert_eq!(StatusCode::Conflict, error.unwrap());
 
     // Working Case (overwrite=true)
-    let overwrite_response = blob_client
-        .upload(
-            RequestContent::from(new_data.to_vec()),
-            true,
-            u64::try_from(new_data.len())?,
-            None,
-        )
+    blob_client
+        .upload(RequestContent::from(new_data.to_vec()), None)
         .await?;
     let response = blob_client.download(None).await?;
     let content_length = response.content_length()?;
 
     // Assert
-    assert_eq!(overwrite_response.status(), StatusCode::Created);
     let (status_code, _, response_body) = response.deconstruct();
     assert!(status_code.is_success());
     assert_eq!(29, content_length.unwrap());
@@ -247,12 +234,7 @@ async fn test_download_blob(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     let data = b"hello rusty world";
 
     blob_client
-        .upload(
-            RequestContent::from(data.to_vec()),
-            false,
-            u64::try_from(data.len())?,
-            None,
-        )
+        .upload(RequestContent::from(data.to_vec()), None)
         .await?;
     let response = blob_client.download(None).await?;
 
@@ -288,12 +270,7 @@ async fn test_set_blob_metadata(ctx: TestContext) -> Result<(), Box<dyn Error>> 
         ..Default::default()
     };
     blob_client
-        .upload(
-            RequestContent::from(data.to_vec()),
-            false,
-            u64::try_from(data.len())?,
-            Some(options_with_metadata),
-        )
+        .upload(RequestContent::from(data.to_vec()), Some(options_with_metadata))
         .await?;
     // Assert
     let response = blob_client.get_properties(None).await?;
@@ -465,12 +442,7 @@ async fn test_leased_blob_operations(ctx: TestContext) -> Result<(), Box<dyn Err
         ..Default::default()
     };
     blob_client
-        .upload(
-            RequestContent::from(data.to_vec()),
-            true,
-            u64::try_from(data.len())?,
-            Some(upload_options),
-        )
+        .upload(RequestContent::from(data.to_vec()), Some(upload_options))
         .await?;
 
     // Assert
@@ -620,12 +592,7 @@ async fn test_encoding_edge_cases(ctx: TestContext) -> Result<(), Box<dyn Error>
 
         // Upload Blob
         blob_client_new
-            .upload(
-                RequestContent::from(b"hello rusty world".to_vec()),
-                true,
-                17,
-                None,
-            )
+            .upload(RequestContent::from(b"hello rusty world".to_vec()), None)
             .await?;
 
         // Get Properties
@@ -648,12 +615,7 @@ async fn test_encoding_edge_cases(ctx: TestContext) -> Result<(), Box<dyn Error>
 
         // Upload Blob
         blob_client_from_url
-            .upload(
-                RequestContent::from(b"hello rusty world".to_vec()),
-                true,
-                17,
-                None,
-            )
+            .upload(RequestContent::from(b"hello rusty world".to_vec()), None)
             .await?;
 
         // Get Properties
@@ -665,12 +627,7 @@ async fn test_encoding_edge_cases(ctx: TestContext) -> Result<(), Box<dyn Error>
 
         // Upload Blob
         blob_client_from_cc
-            .upload(
-                RequestContent::from(b"hello rusty world".to_vec()),
-                true,
-                17,
-                None,
-            )
+            .upload(RequestContent::from(b"hello rusty world".to_vec()), None)
             .await?;
 
         // Get Properties
@@ -972,12 +929,7 @@ async fn test_managed_download(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     let data: [u8; DATA_LEN] = recording.random();
 
     blob_client
-        .upload(
-            RequestContent::from(data.to_vec()),
-            false,
-            u64::try_from(data.len())?,
-            None,
-        )
+        .upload(RequestContent::from(data.to_vec()), None)
         .await?;
 
     for (parallel, partition_len, download_range, expected_gets) in [
@@ -1037,7 +989,7 @@ async fn test_managed_download_empty(ctx: TestContext) -> Result<(), Box<dyn Err
     let blob_client = container_client.blob_client(&get_blob_name(recording));
 
     blob_client
-        .upload(RequestContent::from(vec![]), false, 0, None)
+        .upload(RequestContent::from(vec![]), None)
         .await?;
 
     request_count.store(0, Ordering::Relaxed);